refactor/rename project to bank name
