+++ app/src/androidTest/java/com/example/cjj/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.emptyactivity
+package com.example.cjj
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
