+++ app/src/main/java/com/example/cjj/LoginPage.kt
@@ -1,4 +1,4 @@
-package com.example.emptyactivity
+package com.example.cjj
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -6,10 +6,16 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
+import androidx.compose.material.OutlinedTextField
 import androidx.compose.material3.Button
+import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.semantics.onClick
@@ -18,11 +24,11 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 
-
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun SignupPage(
+fun LoginPage(
     onSuccess: () -> Unit = {},
-    onClickLogin: () -> Unit = {},
+    onClickSignup: () -> Unit = {},
     modifier: Modifier = Modifier
 ) {
     Column(
@@ -31,7 +37,7 @@ fun SignupPage(
         verticalArrangement = Arrangement.Center,
     ) {
         Text(
-            text = "Create an Account",
+            text = "Log In",
             style = MaterialTheme.typography.headlineMedium,
             fontWeight = FontWeight.Bold,
         )
@@ -45,46 +51,70 @@ fun SignupPage(
             placeholder = "password",
             onValueChange = {},
         )
-        LoginSignupTextField(
-            label = "Repeat Password",
-            placeholder = "password",
-            onValueChange = {},
-        )
         Button(
             modifier = Modifier
                 .padding(10.dp)
                 .width(200.dp)
                 .height(50.dp)
                 .semantics {
-                    onClick(label = "sign up for an account", action = null)
+                    onClick(label = "log in to your account", action = null)
                 },
             onClick = onSuccess
         ) {
             Text(
-                text = "Sign Up",
+                text = "Log in",
                 style = MaterialTheme.typography.titleLarge
             )
         }
         Button(
             modifier = Modifier
                 .padding(10.dp)
-                .width(300.dp)
+                .width(240.dp)
                 .height(50.dp)
                 .semantics {
-                    onClick(label = "return to the log in screen", action = null)
+                    onClick(label = "create an account", action = null)
                 },
-            onClick = onClickLogin
+            onClick = onClickSignup
         ) {
             Text(
-                text = "Log in with an existing account",
-                style = MaterialTheme.typography.titleMedium
+                text = "Create an account",
+                style = MaterialTheme.typography.titleLarge
             )
         }
     }
 }
 
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+        /**
+         * A text field for a user's email, password, etc. As the name implies, it should only be used for the
+         * login and signup pages.
+         * @param placeholder The text field's placeholder text.
+         * @param onValueChange The function that will be called with the new value passed in after
+         * the text field's value gets changed.
+         */
+fun LoginSignupTextField(label: String, placeholder: String, onValueChange: (newValue: String) -> Unit, modifier: Modifier = Modifier) {
+    var inputText by remember { mutableStateOf("") }
+
+    OutlinedTextField(
+        singleLine = true,
+        label = { Text(label) },
+        value = inputText,
+        onValueChange = { inputText = it; onValueChange(inputText) },
+        placeholder = @Composable {
+            Text(
+                text = placeholder
+            )
+        },
+        modifier = modifier
+            .width(300.dp)
+            .padding(5.dp)
+    )
+}
+
 @Preview
 @Composable
-fun SignupPagePreview(modifier: Modifier = Modifier) {
-    SignupPage(modifier = modifier)
+fun LoginPagePreview(modifier: Modifier = Modifier) {
+    LoginPage(modifier = modifier)
 }
