+++ app/src/main/java/com/example/cjj/data/BankData.kt
@@ -1,4 +1,4 @@
-package com.example.emptyactivity.data
+package com.example.cjj.data
 
 enum class AccountType
 {
