+++ app/src/main/java/com/example/cjj/ui/theme/Shape.kt
@@ -1,4 +1,4 @@
-package com.example.emptyactivity.ui.theme
+package com.example.cjj.ui.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Shapes
